Abstracting making trades in interactive hyperdrive
